Allows dotnet to build when testing and packaging
It seems that building and packaging in two separate tasks
leads to the creation of an empty Nuget package, at least on OS X.

So, we let the dotnet CLI tool build from sources before packaging
as well as before running the tests. It would have been nice to handle
these dependencies with Cake and skip building using the '--no-build'
flag but dotnet doesn't seem to work reliably enough.