Adds a shim for nuget.exe that runs it on Mono
At this time, Cake isn't able to install tools and addins
from NuGet when running on Unix. The reason is that Cake starts
a process running 'nuget.exe' in order to install the necessary packages.
This solution works fine on Windows but it doesn't work at all on Unix,
where the process simply exits with 'error 13'.

The workaround is to create a shim for 'nuget.exe' in the 'tools'
directory where Cake will look for it; the shim, then, runs
the real 'nuget.exe' through Mono and everything is fine.